Refactor: Decouple Queue Infrastructure from Job Handler Logic
